System Log Collection
Phosphor Debug Collector already provides mechanisms to collect debug logs and system parameters, however it does not have the support for collection of system information log with password protection. Also, there is no provisioning to downloading of generated log file using webUI. So we are adding the new plugins for the system debug collections with password protection.